Refactored fake SPI, add get_daten, get_status, draw graph in qt thread

+++ Raspberry/daten.h
@@ -1,12 +1,13 @@
 #ifndef DATEN_H
 #define DATEN_H
+#include <stdint.h>
 
 class Daten
 {
 public:
-    Daten(float n_zeitpunkt, float n_messwert):zeitpunkt(n_zeitpunkt), messwert(n_messwert) {};
+    Daten(uint32_t n_zeitpunkt, float n_messwert):zeitpunkt(n_zeitpunkt), messwert(n_messwert) {};
     virtual ~Daten(){}
-    const float zeitpunkt;
+    const uint32_t zeitpunkt;
     const float messwert;
 };
